backend/requirements.txt
flask
gunicorn
flask-cors
torch
timm==0.4.12
opencv-python
Pillow
torchvision
werkzeug
